* Follow up to bd2503e and 7355677 -- Modified field data of `equip_locations` to mediumint -- Updated converter tool for item_db following this update -- Updated item_db.sql, item_db2.sql, item_db_re.sql, item_db_re2.sql -- CAUTION: Please update item_db MySQL tables by using 'upgrade_20131115_2.sql'

+++ sql-files/item_db2.sql
@@ -18,7 +18,7 @@ CREATE TABLE `item_db2` (
   `equip_jobs` int(12) unsigned default NULL,
   `equip_upper` tinyint(8) unsigned default NULL,
   `equip_genders` tinyint(2) unsigned default NULL,
-  `equip_locations` smallint(4) unsigned default NULL,
+  `equip_locations` mediumint(7) unsigned default NULL,
   `weapon_level` tinyint(2) unsigned default NULL,
   `equip_level` tinyint(3) unsigned default NULL,
   `refineable` tinyint(1) unsigned default NULL,
